requirements.txt
apache-airflow-providers-databricks[all]==7.4.0
python-dotenv==1.1.0
flask_appbuilder>=4.3.6

# Azure SDKs
azure-core==1.34.0
azure-identity==1.23.0
azure-storage-blob==12.25.1
azure-storage-file-datalake==12.20.0
msal==1.32.3
msal-extensions==1.3.1

certifi==2025.4.26
cryptography==44.0.3
